hospital_assistant/requirements.txt
streamlit
nltk
langchain
Ollama
python-dotenv
langchain-ollama

transformers
datasets
peft
trl
accelerate
bitsandbytes
